added extra parameter in hash function
I have added an extra parameter to the public key hash calculation to act as a linkability limiter. I have run all tests with this parameter and everyone passes. Now I need to change the code such that if the challenge is the same but the extra parameter is different then the linkability is false. Since one can sign the same message again, if one is in a different situation / setting.